more JSON support

+++ src/controllers/loansController.ts
@@ -1,8 +1,7 @@
-import { type Request, type Response, type NextFunction } from 'express';
+import { type NextFunction, type Request, type Response } from 'express';
 import Loan, { type ILoan } from '../models/loanModel';
-import { realValidateUser } from './validateUser';
 import { AccountLevel } from '../models/userModel';
-
+import { isAdmin, realValidateUser } from './validateUser';
 async function getLoans(): Promise<ILoan[]> {
   return await Loan.find({});
 }
@@ -13,8 +12,15 @@ export const getLoanJSON = async (
   next: NextFunction
 ): Promise<void> => {
   try {
-    const loans = await getLoans();
-    res.json({ loans });
+    const token = req.cookies.token as string;
+    const user = await realValidateUser(token);
+    if (user?.user !== undefined) {
+      const loan = await Loan.findById(req.body.id as string);
+
+      res.json({ loan: [loan] });
+    } else {
+      res.redirect('/loans');
+    }
   } catch (err) {
     next(err);
   }
@@ -26,8 +32,25 @@ export const getLoansJSON = async (
   next: NextFunction
 ): Promise<void> => {
   try {
-    const loans = await getLoans();
-    res.json({ loans });
+    const token = req.headers.token as string;
+    const user = await realValidateUser(token);
+    console.log(user);
+
+    if (user?.accountLevel === AccountLevel.Admin) {
+      const loans = await getLoans();
+
+      res.json({ loans });
+    } else if (user?.user !== undefined) {
+      //   const id = new ObjectId(user.user._id);
+
+      const loans = await Loan.findOne({
+        user: user.user._id
+      });
+
+      res.json({ loans: [loans] });
+    } else {
+      throw new Error("User either doesn't exist, or isn't and admin");
+    }
   } catch (err) {
     next(err);
   }
@@ -47,8 +70,13 @@ export const getLoansWeb = async (
 
       res.render('loans/index', { loans });
     } else if (user?.user !== undefined) {
-      const loans = Loan.find({ user: user.user._id });
-      res.render('loans/index', { loans });
+      //   const id = new ObjectId(user.user._id);
+
+      const loans = await Loan.findOne({
+        user: user.user._id
+      });
+
+      res.render('loans/index', { loans: [loans] });
     } else {
       throw new Error("User either doesn't exist, or isn't and admin");
     }
@@ -79,14 +107,17 @@ export const createLoanWeb = async (
   res: Response,
   next: NextFunction
 ): Promise<void> => {
+  console.log('asdfasdfasdfwerqwe');
+
   try {
-    const { amount, interestRate, term, borrower } = req.body;
+    const { amount, interestRate, term, user } = req.body;
     const loan = new Loan({
       amount,
       interestRate,
       term,
-      borrower
+      user
     });
+
     await loan.save();
     res.redirect('/loans');
   } catch (err) {
@@ -121,11 +152,38 @@ export const deleteLoanWeb = async (
   next: NextFunction
 ): Promise<Response<any, Record<string, any>> | undefined> => {
   try {
-    const loan = await Loan.findByIdAndDelete(req.params.id);
-    if (loan == null) {
-      return res.status(404).send('Loan not found');
+    const userIsAdmin = await isAdmin(req.cookies.token as string);
+    if (userIsAdmin) {
+      const loan = await Loan.findByIdAndDelete(req.params.id);
+      if (loan == null) {
+        return res.status(404).send('Loan not found');
+      } else {
+        throw new Error('User either does not exist, or is not Admin');
+      }
     }
     res.redirect('/loans');
+  } catch (err) {
+    next(err);
+  }
+};
+
+export const deleteLoanJson = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+): Promise<Response<any, Record<string, any>> | undefined> => {
+  try {
+    const userIsAdmin = await isAdmin(req.headers.token as string);
+    if (userIsAdmin) {
+      const loan = await Loan.findByIdAndDelete(req.params.id);
+      if (loan == null) {
+        return res.status(404).json('Loan not found');
+      } else {
+        res.json(`Successfully deleted loan: ${loan?._id as string}`);
+      }
+    } else {
+      throw new Error('User either does not exist, or is not Admin');
+    }
   } catch (err) {
     next(err);
   }
@@ -137,8 +195,6 @@ export const editLoanWeb = async (
   next: NextFunction
 ): Promise<void> => {
   try {
-    console.log('asdiofjaiosdfjoi');
-
     const loan = await Loan.findById(req.params.id);
     res.render('loans/edit', { loan });
   } catch (err) {
